update tests with send_request/accept_friend

+++ app/controllers/friendships_controller.rb
@@ -4,8 +4,8 @@
 class FriendshipsController < ApplicationController
   def create
     if (request = Request.find_by(sender_id: params[:friend_id], receiver_id: current_user.id))
-      current_user.passive_friendships.create(active_friend_id: params[:friend_id])
-      request.destroy
+      friendship = current_user.accept_friend(params[:friend_id])
+      request.destroy if friendship.save
     end
     redirect_back(fallback_location: requests_path)
   end
